requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
channels==3.0.3
channels-redis==4.2.0
daphne==3.0.2
dj-database-url==2.2.0
Django==4.2
django-bootstrap4==24.3
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==23.0.0
pillow==9.0.0
pytz==2022.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
redis==5.0.7
uvicorn==0.30.6
whitenoise==6.7.0
wheel>=0.38.4